Move Quoted pretty printing to its own module
